Remove unnecessary `str_replace()` to fix notices.
This line throws a notice because it's treating an array as a string. `str_replace()` can accept an array for `$subject`, but it will not recurse through it. The notice gets thrown when an element of `$tag` is also an array, like with `$tag['types']`.

It looks like the underlying problem was with the inline docs, and they have since been fixed to avoid the need for this statement.

See https://core.trac.wordpress.org/ticket/28298
See https://meta.trac.wordpress.org/ticket/492
See https://meta.trac.wordpress.org/ticket/1042