Switch to lodash throttle
In several browsers there is a delay before the scroll position is restored.
This results in the ToC not being repositioned if the page is refreshed when scrolled to the bottom.
Lodash Throttle provides support for `trailing` events to always be fired, which fixes the issue, whereas rafThrottle does not.