Update Composer packages causing failures.
This updates the `wporg/wporg-mu-plugins` and `wporg/wporg-parent-2021` dependencies in Composer.

When running `yarn setup:tools` for the first time, the command will fail due to the pinned commit hashes being removed from the repository.

It seems like this can happen through rebases, but it's not clear why this is happening for these two repositories.